VimeoPlayer API improvements
