+++ Assets/TinyPlanetRecorder.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: d1c886ab30faf4148b89504b9e1d1789
+guid: 6e6ef09d6a5af064aa500c30316b1fd2
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
