+++ Assets/Vimeo/Models/Materials.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 060c83879751cf44cada88b3d3236dc2
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
